requirements.txt
# Market Analysis Agent - Dependencies
aiohttp>=3.9.0
python-dotenv>=1.0.0
yfinance>=0.2.0
